Revising the "discrete" module and implemented memory peripheral.

+++ galaktec-digital-cpu-peripheral/src/address_map.rs
@@ -1,11 +1,11 @@
 #[derive(Debug, Copy, Clone)]
 pub struct AddressMap {
-    base_address: u32,
-    block_size: u32,
+    base_address: usize,
+    block_size: usize,
 }
 
 impl AddressMap {
-    pub fn new(base_address: u32, block_size: u32) -> Self {
+    pub fn new(base_address: usize, block_size: usize) -> Self {
         debug_assert_ne!(block_size, 0, "Block size can not be zero");
 
         AddressMap {
@@ -14,15 +14,15 @@ impl AddressMap {
         }
     }
 
-    fn end_address(&self) -> u32 {
+    fn end_address(&self) -> usize {
         self.base_address + self.block_size
     }
 
-    pub fn address_in_range(&self, address: u32) -> bool {
+    pub fn address_in_range(&self, address: usize) -> bool {
         self.base_address <= address && address < self.end_address()
     }
 
-    pub fn relative_address(&self, address: u32) -> u32 {
+    pub fn relative_address(&self, address: usize) -> usize {
         debug_assert!(self.address_in_range(address), "Address not in range");
         address - self.base_address
     }
